refactor(Vuex): Refactor class update
Change to one-at-a-time add-or-update mutation

BREAKING CHANGES
    dto folder renamed api
    data.js renamed kentico.js
    Types.mutations.CLEAR_CLASSES_BY_CLASSIFICATION,
    Types.mutations.ADD_CLASS and
    Types.mutations.CLEAR_CLASSES_BY_CODENAME
    removed
